+++ exporters/otlp/logs/src/testGrpcNetty/java/io/opentelemetry/exporter/otlp/logs/ExportTest.java
@@ -23,11 +23,11 @@ import io.opentelemetry.proto.collector.logs.v1.ExportLogsServiceResponse;
 import io.opentelemetry.proto.collector.logs.v1.LogsServiceGrpc;
 import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
 import io.opentelemetry.sdk.logs.data.LogData;
-import io.opentelemetry.sdk.logs.data.LogRecord;
 import io.opentelemetry.sdk.logs.data.Severity;
 import io.opentelemetry.sdk.resources.Resource;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
+import java.time.Instant;
 import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
@@ -39,10 +39,10 @@ class ExportTest {
 
   private static final List<LogData> LOGS =
       Collections.singletonList(
-          LogRecord.builder(
+          LogData.builder(
                   Resource.create(Attributes.builder().put("testKey", "testValue").build()),
                   InstrumentationLibraryInfo.create("instrumentation", "1"))
-              .setEpochMillis(System.currentTimeMillis())
+              .setEpoch(Instant.now())
               .setTraceId(TraceId.getInvalid())
               .setSpanId(SpanId.getInvalid())
               .setFlags(TraceFlags.getDefault().asByte())