+++ exporters/otlp/logs/src/testOkHttpOnly/java/io/opentelemetry/exporter/otlp/logs/OkHttpOnlyExportTest.java
@@ -20,12 +20,12 @@ import io.opentelemetry.exporter.otlp.internal.grpc.OkHttpGrpcExporterBuilder;
 import io.opentelemetry.proto.collector.logs.v1.ExportLogsServiceResponse;
 import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
 import io.opentelemetry.sdk.logs.data.LogData;
-import io.opentelemetry.sdk.logs.data.LogRecord;
 import io.opentelemetry.sdk.logs.data.Severity;
 import io.opentelemetry.sdk.resources.Resource;
 import java.net.InetAddress;
 import java.net.UnknownHostException;
 import java.nio.charset.StandardCharsets;
+import java.time.Instant;
 import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
@@ -40,10 +40,10 @@ class OkHttpOnlyExportTest {
 
   private static final List<LogData> LOGS =
       Collections.singletonList(
-          LogRecord.builder(
+          LogData.builder(
                   Resource.create(Attributes.builder().put("testKey", "testValue").build()),
                   InstrumentationLibraryInfo.create("instrumentation", "1"))
-              .setEpochMillis(System.currentTimeMillis())
+              .setEpoch(Instant.now())
               .setTraceId(TraceId.getInvalid())
               .setSpanId(SpanId.getInvalid())
               .setFlags(TraceFlags.getDefault().asByte())