+++ integration-tests/src/testOtlpCommon/java/io/opentelemetry/integrationtest/OtlpExporterIntegrationTest.java
@@ -52,7 +52,7 @@ import io.opentelemetry.proto.trace.v1.ResourceSpans;
 import io.opentelemetry.proto.trace.v1.Span.Link;
 import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
 import io.opentelemetry.sdk.logs.data.Body;
-import io.opentelemetry.sdk.logs.data.LogRecord;
+import io.opentelemetry.sdk.logs.data.LogData;
 import io.opentelemetry.sdk.logs.data.Severity;
 import io.opentelemetry.sdk.logs.export.LogExporter;
 import io.opentelemetry.sdk.metrics.SdkMeterProvider;
@@ -72,7 +72,6 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.CompletionStage;
-import java.util.concurrent.TimeUnit;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import org.junit.jupiter.api.AfterAll;
@@ -375,8 +374,8 @@ abstract class OtlpExporterIntegrationTest {
   }
 
   private static void testLogExporter(LogExporter logExporter) {
-    LogRecord logRecord =
-        LogRecord.builder(
+    LogData logData =
+        LogData.builder(
                 RESOURCE,
                 InstrumentationLibraryInfo.create(
                     OtlpExporterIntegrationTest.class.getName(), null))
@@ -387,11 +386,11 @@ abstract class OtlpExporterIntegrationTest {
             .setSeverityText("DEBUG")
             .setTraceId(IdGenerator.random().generateTraceId())
             .setSpanId(IdGenerator.random().generateSpanId())
-            .setEpochNanos(TimeUnit.MILLISECONDS.toNanos(Instant.now().toEpochMilli()))
+            .setEpoch(Instant.now())
             .setFlags(0)
             .build();
 
-    logExporter.export(Collections.singletonList(logRecord));
+    logExporter.export(Collections.singletonList(logData));
 
     await()
         .atMost(Duration.ofSeconds(30))
@@ -425,14 +424,13 @@ abstract class OtlpExporterIntegrationTest {
                     .setValue(AnyValue.newBuilder().setStringValue("value").build())
                     .build()));
     assertThat(protoLog.getSeverityNumber().getNumber())
-        .isEqualTo(logRecord.getSeverity().getSeverityNumber());
+        .isEqualTo(logData.getSeverity().getSeverityNumber());
     assertThat(protoLog.getSeverityText()).isEqualTo("DEBUG");
     assertThat(TraceId.fromBytes(protoLog.getTraceId().toByteArray()))
-        .isEqualTo(logRecord.getTraceId());
-    assertThat(SpanId.fromBytes(protoLog.getSpanId().toByteArray()))
-        .isEqualTo(logRecord.getSpanId());
-    assertThat(protoLog.getTimeUnixNano()).isEqualTo(logRecord.getEpochNanos());
-    assertThat(protoLog.getFlags()).isEqualTo(logRecord.getFlags());
+        .isEqualTo(logData.getTraceId());
+    assertThat(SpanId.fromBytes(protoLog.getSpanId().toByteArray())).isEqualTo(logData.getSpanId());
+    assertThat(protoLog.getTimeUnixNano()).isEqualTo(logData.getEpochNanos());
+    assertThat(protoLog.getFlags()).isEqualTo(logData.getFlags());
   }
 
   private static class OtlpGrpcServer extends ServerExtension {