+++ sdk/logs/src/main/java/io/opentelemetry/sdk/logs/data/LogDataBuilder.java
@@ -9,11 +9,13 @@ import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.api.common.AttributesBuilder;
 import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
 import io.opentelemetry.sdk.resources.Resource;
+import java.time.Instant;
 import java.util.concurrent.TimeUnit;
 import javax.annotation.Nullable;
 
-/** Builder for {@link LogRecord}. */
-public final class LogRecordBuilder {
+/** Builder for {@link LogData}. */
+public final class LogDataBuilder {
+
   private final Resource resource;
   private final InstrumentationLibraryInfo instrumentationLibraryInfo;
 
@@ -27,74 +29,82 @@ public final class LogRecordBuilder {
   private Body body = Body.stringBody("");
   private final AttributesBuilder attributeBuilder = Attributes.builder();
 
-  LogRecordBuilder(Resource resource, InstrumentationLibraryInfo instrumentationLibraryInfo) {
+  LogDataBuilder(Resource resource, InstrumentationLibraryInfo instrumentationLibraryInfo) {
     this.resource = resource;
     this.instrumentationLibraryInfo = instrumentationLibraryInfo;
   }
 
-  public LogRecordBuilder setEpochNanos(long timestamp) {
-    this.epochNanos = timestamp;
+  /** Set the epoch timestamp using the timestamp and unit. */
+  public LogDataBuilder setEpoch(long timestamp, TimeUnit unit) {
+    this.epochNanos = unit.toNanos(timestamp);
     return this;
   }
 
-  public LogRecordBuilder setEpochMillis(long timestamp) {
-    return setEpochNanos(TimeUnit.MILLISECONDS.toNanos(timestamp));
+  /** Set the epoch timestamp using the instant. */
+  public LogDataBuilder setEpoch(Instant instant) {
+    this.epochNanos = TimeUnit.SECONDS.toNanos(instant.getEpochSecond()) + instant.getNano();
+    return this;
   }
 
-  public LogRecordBuilder setTraceId(String traceId) {
+  /** Set the trace id. */
+  public LogDataBuilder setTraceId(String traceId) {
     this.traceId = traceId;
     return this;
   }
 
-  public LogRecordBuilder setSpanId(String spanId) {
+  /** Set the span id. */
+  public LogDataBuilder setSpanId(String spanId) {
     this.spanId = spanId;
     return this;
   }
 
-  public LogRecordBuilder setFlags(int flags) {
+  /** Set the flags. */
+  public LogDataBuilder setFlags(int flags) {
     this.flags = flags;
     return this;
   }
 
-  public LogRecordBuilder setSeverity(Severity severity) {
+  /** Set the severity. */
+  public LogDataBuilder setSeverity(Severity severity) {
     this.severity = severity;
     return this;
   }
 
-  public LogRecordBuilder setSeverityText(String severityText) {
+  /** Set the severity text. */
+  public LogDataBuilder setSeverityText(String severityText) {
     this.severityText = severityText;
     return this;
   }
 
-  public LogRecordBuilder setName(String name) {
+  /** Set the name. */
+  public LogDataBuilder setName(String name) {
     this.name = name;
     return this;
   }
 
-  public LogRecordBuilder setBody(Body body) {
+  /** Set the body. */
+  public LogDataBuilder setBody(Body body) {
     this.body = body;
     return this;
   }
 
-  public LogRecordBuilder setBody(String body) {
+  /** Set the body string. */
+  public LogDataBuilder setBody(String body) {
     return setBody(Body.stringBody(body));
   }
 
-  public LogRecordBuilder setAttributes(Attributes attributes) {
+  /** Set the attributes. */
+  public LogDataBuilder setAttributes(Attributes attributes) {
     this.attributeBuilder.putAll(attributes);
     return this;
   }
 
-  /**
-   * Build a LogRecord instance.
-   *
-   * @return value object being built
-   */
-  public LogRecord build() {
+  /** Build a {@link LogData} instance from the configured properties. */
+  public LogData build() {
     if (epochNanos == 0) {
       epochNanos = TimeUnit.MILLISECONDS.toNanos(System.currentTimeMillis());
     }
-    return LogRecord.create(
+    return LogDataImpl.create(
         resource,
         instrumentationLibraryInfo,
         epochNanos,