+++ sdk/logs/src/test/java/io/opentelemetry/sdk/logs/util/TestUtil.java
@@ -10,17 +10,18 @@ import io.opentelemetry.api.trace.SpanId;
 import io.opentelemetry.api.trace.TraceFlags;
 import io.opentelemetry.api.trace.TraceId;
 import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
-import io.opentelemetry.sdk.logs.data.LogRecord;
+import io.opentelemetry.sdk.logs.data.LogData;
 import io.opentelemetry.sdk.logs.data.Severity;
 import io.opentelemetry.sdk.resources.Resource;
+import java.util.concurrent.TimeUnit;
 
 public final class TestUtil {
 
-  public static LogRecord createLog(Severity severity, String message) {
-    return LogRecord.builder(
+  public static LogData createLogData(Severity severity, String message) {
+    return LogData.builder(
             Resource.create(Attributes.builder().put("testKey", "testValue").build()),
             InstrumentationLibraryInfo.create("instrumentation", "1"))
-        .setEpochMillis(System.currentTimeMillis())
+        .setEpoch(System.currentTimeMillis(), TimeUnit.MILLISECONDS)
         .setTraceId(TraceId.getInvalid())
         .setSpanId(SpanId.getInvalid())
         .setFlags(TraceFlags.getDefault().asByte())